+++ Changelog.md
@@ -13,6 +13,8 @@
 - Removed the Vanilo v2 `Framework` namespace compatibility layer
 - Added the `currency` field to the orders table
 - Added `currency` field to the Channel model/table
+- Added the `payment_method_id` to the orders table
+- Added the processing of the `payment_method_id` field to the OrderFactory (Foundation)
 - Added the `is_hidden` field to the `Property` model
 - Added the `withImages` and `withChannels` methods to the product search class (eager loads media)
 - Added the `channelables` table for being many-to-many polymorphic relationships with channels and arbitrary models